+++ realm/realm-library/src/androidTest/kotlin/io/realm/UnmanagedSetTester.kt
@@ -148,6 +148,10 @@ class UnmanagedSetTester<T : Any>(
 
     override fun insertList() = Unit  // Not applicable
 
+    override fun insertOrUpdate() = Unit  // Not applicable
+
+    override fun insertOrUpdateList() = Unit  // Not applicable
+
     override fun copyToRealm() = Unit  // Not applicable
 
     override fun copyToRealmOrUpdate() = Unit  // Not applicable